trippo_user/scripts/requirements.txt
firebase-admin>=6.0.0
google-cloud-firestore>=2.11.0

